feat: implement basic project structure
