+++ binding.gyp
@@ -1,0 +1,7 @@
+{
+  "targets": [
+    {
+      "target_name": "binding"
+    }
+  ]
+}
